Update MarketTable and Sidebar file structure
