+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/OutcomeSelection.js
