+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/OutcomeSelection.scss
