+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/Probabilities.js
