+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/ProbabilityWording.js
