+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/ResolutionDate.js
