+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/ToggleConditional.js
