+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/index.js
