+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/marketRow.scss
