+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/probabilities.scss
