+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/probabilityWording.scss
