+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/resolutionDate.scss
