+++ app/src/MarketTable/CategoricalMarketTable/MarketRow/toggleConditional.scss
