+++ app/src/MarketTable/CategoricalMarketTable/marketTable.scss
