+++ app/src/MarketTable/ScalarMarketTable/Graph.js
