+++ app/src/MarketTable/ScalarMarketTable/Graph.scss
