+++ app/src/MarketTable/ScalarMarketTable/ResolutionTime.js
