+++ app/src/MarketTable/ScalarMarketTable/marketTable.scss
