+++ app/src/MarketTable/ScalarMarketTable/utils/prepareQueryData.js
