+++ app/src/Sidebar/CategoricalSidebar/Balance/balance.scss
