+++ app/src/Sidebar/CategoricalSidebar/Buy/buy.scss
