+++ app/src/Sidebar/CategoricalSidebar/Positions/Sell.js
