+++ app/src/Sidebar/CategoricalSidebar/Positions/positions.scss
