+++ app/src/Sidebar/CategoricalSidebar/sidebar.scss
@@ -1,4 +1,4 @@
-@import "../scss/_variables.scss";
+@import "~scss/_variables.scss";
 
 .sidebar {
   height: 100%;
