+++ app/src/Sidebar/ScalarSidebar/Buy/buy.scss
