+++ app/src/Sidebar/ScalarSidebar/Sell/positions.scss
