+++ app/src/Sidebar/ScalarSidebar/index.js
