+++ app/src/Sidebar/ScalarSidebar/sidebar.scss
